% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistnet.R
\name{assistnet}
\alias{assistnet}
\title{Calculate table of assists and some useful player statistics}
\usage{
assistnet(data, assist = "assist", player = "player",
  points = "points", event.type = "event_type")
}
\arguments{
\item{data}{A play-by-play dataframe}

\item{assist}{A string for the name of the column with players who made the assists}

\item{player}{A string for the name of the column with players responsible for the recorded events}

\item{points}{A string for the name of the column with points made}

\item{event.type}{A string for the name of the column with type of event}
}
\value{
A list with 3 elements: assistTable (a table), nodeStats (a dataframe), and a net (a network object)
}
\description{
Calculate table of assists and some useful player statistics
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP.GSW <- subset(PbP, team=="GSW" & player!="")
out <- assistnet(PbP.GSW)
plot(out, layout="circle", edge.thr=30, node.col="FGM_ASTp", node.size="ASTPTS")
}
