% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubbleplot.R
\name{bubbleplot}
\alias{bubbleplot}
\title{Create a bubble plot}
\usage{
bubbleplot(data, id, x, y, col, size, scale.size = TRUE, labels = NULL,
  mx = NULL, my = NULL, mcol = NULL, title = NULL, repel = TRUE)
}
\arguments{
\item{data}{A dataframe}

\item{id}{The name of the ID variable}

\item{x}{Name of variable on the x axis}

\item{y}{Name of variable on the y axis}

\item{col}{Name of variable on the color axis}

\item{size}{Name of variable on the size axis}

\item{scale.size}{If TRUE, 'size' variable is rescaled between 0 and 100.}

\item{labels}{Vector of labels for 4 variables}

\item{mx}{mx}

\item{my}{my}

\item{mcol}{mcol}

\item{title}{Title for the plot}

\item{repel}{Activate text repelling}
}
\value{
A ggplot2 object
}
\description{
Create a bubble plot
}
