% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_plot_mixed.R
\name{corr_plot_mixed}
\alias{corr_plot_mixed}
\title{Plot inequality}
\usage{
corr_plot_mixed(cor_mtx, cor_mtest, sl)
}
\arguments{
\item{cor_mtx}{cor_mtx}

\item{cor_mtest}{cor_mtest}

\item{sl}{sl}
}
\value{
Nothing
}
\description{
Plot inequality
}
