% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corranalysis.R
\name{corranalysis}
\alias{corranalysis}
\title{Correlation analysis}
\usage{
corranalysis(data, threshold = 0, sl = 0.05)
}
\arguments{
\item{data}{A dataframe}

\item{threshold}{threshold}

\item{sl}{Significance level}
}
\value{
A list with the following elements: corr.mtx, corr.mtx.trunc, cor.mtest, threshold, siglevel
}
\description{
Correlation analysis
}
\examples{
data <- data.frame(Pbox$PTS,Pbox$P3M,Pbox$P2M,
                   Pbox$OREB + Pbox$DREB,Pbox$AST,
                   Pbox$TOV,Pbox$STL,Pbox$BLK)/Pbox$MIN
names(data) <- c("PTS","P3M","P2M","REB","AST","TOV","STL","BLK")
data <- subset(data, Pbox$MIN >= 500)
out <- corranalysis(data, threshold = 0.5)
plot(out)
}
