% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epoints.R
\name{Epoints}
\alias{Epoints}
\title{Plot expected points of player shots as a function of the distance from the basket}
\usage{
Epoints(data, players = NULL, bw = 10, palette = gg_color_hue,
  col.team = "gray", col.hline = "black", xlab = "Shot distance",
  title = NULL, legend = TRUE, var = "shot_distance")
}
\arguments{
\item{data}{A play-by-play data frame}

\item{players}{Subset of players to be displayed}

\item{bw}{A numerical value for the smoothing bandwidth of the kernel density estimator (see \link[stats]{ksmooth})}

\item{palette}{Color palette}

\item{col.team}{Color of the scoring probability line for team}

\item{col.hline}{Color of the dashed horizontal line}

\item{xlab}{x-axis label}

\item{title}{Plot title}

\item{legend}{If TRUE, color legend is displayed (only when 'players' is not NULL)}

\item{var}{A character string giving the numerical variable on the x-axis (default "shot_distance").}
}
\value{
A ggplot2 plot
}
\description{
Plot expected points of player shots as a function of the distance from the basket
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP.GSW <- subset(PbP, team=="GSW" & !is.na(shot_distance))
plrys <- c("Stephen Curry","Kevin Durant")
Epoints(data=PbP.GSW, bw=10, players=plrys, col.team='dodgerblue',
        palette=colorRampPalette(c("gray","black")), col.hline="red")
}
