% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourfactorsPlot.R
\name{fourfactorsPlot}
\alias{fourfactorsPlot}
\title{Plot possession, pace, ratings, and Four Factors}
\usage{
fourfactorsPlot(dtset, title = NULL)
}
\arguments{
\item{dtset}{A dataframe with calculated possession, pace, ratings, and Four Factors}

\item{title}{Plot title}
}
\value{
A list of four ggplot2 plots
}
\description{
Plot possession, pace, ratings, and Four Factors
}
\examples{
data("Tbox")
data("Pbox")
selectedTeams <- c(2,6,10,11)
FF <- fourfactors(Tbox, Obox, sel=selectedTeams)
listPlots <- fourfactorsPlot(FF)
library(gridExtra)
grid.arrange(grobs=listPlots, ncol=2)
}
