% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequality.R
\name{inequality}
\alias{inequality}
\title{Inequality analysis}
\usage{
inequality(data, nplayers)
}
\arguments{
\item{data}{A numerical vector}

\item{nplayers}{Number of players to include in the analysis}
}
\value{
A list with the following elements:  Lorenz curve ("Lorenz") and Gini coefficient ("Gini") for a given number of players
}
\description{
Inequality analysis
}
\examples{
Pbox.BN <- subset(Pbox, Team=="Brooklyn Nets")
out <- inequality(Pbox.BN$PTS, nplayers=8)
print(out)
plot(out)
}
