% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.fourfactors.R
\name{is.fourfactors}
\alias{is.fourfactors}
\title{Functions to check if an object is of class 'fourfactors'.}
\usage{
is.fourfactors(x)
}
\arguments{
\item{x}{any R object}
}
\value{
returns TRUE if its argument is of class 'fourfactors' and FALSE otherwise
}
\description{
Functions to check if an object is of class 'fourfactors'.
}
\examples{
data("Tbox")
data("Pbox")
selectedTeams <- c(2,6,10,11)
FF <- fourfactors(Tbox, Obox, sel=selectedTeams)
is.fourfactors(FF)
}
