% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.inequality.R
\name{is.inequality}
\alias{is.inequality}
\title{Functions to check if an object is of class 'inequality'.}
\usage{
is.inequality(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
returns TRUE if its argument is of class \code{inequality} and FALSE otherwise
}
\description{
Functions to check if an object is of class 'inequality'.
}
\examples{
Pbox.BN <- subset(Pbox, Team=="Brooklyn Nets")
out <- inequality(Pbox.BN$PTS, npl=8)
is.inequality(out)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\seealso{
\code{\link{inequality}}
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketball.analyzer.help@gmail.com})
}
