% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.networkdata.R
\name{is.networkdata}
\alias{is.networkdata}
\title{Reports whether x is a networkdata object}
\usage{
is.networkdata(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
returns TRUE if its argument is of class 'networkdata' and FALSE otherwise
}
\description{
Reports whether x is a networkdata object
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP.GSW <- subset(PbP, team=="GSW" & player!="")
out <- netcalc(PbP.GSW)
is.networkdata(out)
}
