% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.shotperformance.R
\name{is.shotperformance}
\alias{is.shotperformance}
\title{Reports whether x is a 'shotperformance' object}
\usage{
is.shotperformance(x)
}
\arguments{
\item{x}{an object to test.}
}
\value{
Returns TRUE if its argument is of class \code{shotperformance} and FALSE otherwise.
}
\description{
Reports whether x is a 'shotperformance' object
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP <- scoredifference(PbP_data = PbP, team_name = "GSW", player_data=Pbox, team_data = Tadd)
PbP <- shotclock(PbP_data = PbP, sec_14_after_oreb = FALSE, team_data = Tadd)
shotperf <- shotperformance(PbP_data = PbP, player_data = Pbox, team_data = Tadd,
                shotclock_interval = c(0, 2) , shot_type = "2P"  )
is.shotperformance(shotperf)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\seealso{
\code{\link{shotperformance}}
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
