% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-NBA1718Players.R
\docType{data}
\name{NBA1718Players}
\alias{NBA1718Players}
\title{Data set of NBA 2017-2018}
\format{A data.frame with 605 rows and 22 variables:
\describe{
  \item{Team}{Teams}
  \item{Player}{Players}
  \item{GP}{GP}
  \item{MIN}{MIN}
  \item{PTS}{PTS}
}}
\source{
\url{https://github.com/sndmrc/BasketAnalyzeR}
}
\usage{
NBA1718Players
}
\description{
A data set of NBA 2017-2018 (Players)
}
\keyword{datasets}
