% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcalc.R
\name{netcalc}
\alias{netcalc}
\title{Calculate table of assists and some useful player statistics}
\usage{
netcalc(data, assist = "assist", player = "player")
}
\arguments{
\item{data}{A dataframe}

\item{assist}{assist}

\item{player}{player}
}
\value{
A list with two elements: assistTable (a table) and stats (a dataframe)
}
\description{
Calculate table of assists and some useful player statistics
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP.GSW <- subset(PbP, team=="GSW" & player!="")
out <- netcalc(PbP.GSW)
plot(out, layout="circle", edge.thr=30)
}
