% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkplot.R
\name{networkplot}
\alias{networkplot}
\title{Create a network plot}
\usage{
networkplot(data, assist = "assist", player = "player",
  layout = "kamadakawai", layout.par = list(), node.data = NULL,
  node.size = NULL, node.col = NULL, node.col.lim = NULL,
  node.pal = colorRampPalette(c("white", "blue", "red")),
  edge.pal = colorRampPalette(c("white", "blue", "red")))
}
\arguments{
\item{data}{A dataframe}

\item{assist}{assist}

\item{player}{player}

\item{layout}{Network vertex layout algorithms}

\item{layout.par}{Paramater for network vertex layout algorithms}

\item{node.data}{Data with variables for node size and node color}

\item{node.size}{A string with the variable name for node size}

\item{node.col}{A string with the variable name for node color}

\item{node.col.lim}{A numeric vector of length two providing limits of the scale for node color}

\item{node.pal}{Palette for node colors}

\item{edge.pal}{Palette for edge colors}
}
\value{
A ggplot2 object
}
\description{
Create a network plot
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP.GSW <- subset(PbP, team=="GSW" & player!="")
networkplot(PbP.GSW, layout="circle")
}
