% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pbox.R
\docType{data}
\name{Pbox}
\alias{Pbox}
\title{Players box scores dataset - NBA 2017-2018}
\format{
A data.frame with 605 rows and 22 variables:
\describe{
\item{Team}{Analyzed team, character}
\item{Player}{Analyzed player, character}
\item{GP}{Games Played, numeric}
\item{MIN}{Minutes Played, numeric}
\item{PTS}{Points Made, numeric}
\item{P2M}{2-Point Field Goals (Made), numeric}
\item{P2A}{2-Point Field Goals (Attempted), numeric}
\item{P2p}{2-Point Field Goals (Percentage), numeric}
\item{P3M}{3-Point Field Goals (Made), numeric}
\item{P3A}{3-Point Field Goals (Attempted), numeric}
\item{P3p}{3-Point Field Goals (Percentage), numeric}
\item{FTM}{Free Throws (Made), numeric}
\item{FTA}{Free Throws (Attempted), numeric}
\item{FTp}{Free Throws (Percentage), numeric}
\item{OREB}{Offensive Rebounds, numeric}
\item{DREB}{Defensive Rebounds, numeric}
\item{AST}{Assists, numeric}
\item{TOV}{Turnovers, numeric}
\item{STL}{Steals, numeric}
\item{BLK}{Blocks, numeric}
\item{PF}{Personal Fouls, numeric}
\item{PM}{Plus/Minus, numeric}
}
}
\usage{
Pbox
}
\description{
In this data frame, cases (rows) are players and variables (columns) are referred to the individual achievements in the NBA 2017-2018 Championship
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
\keyword{datasets}
