% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PbP.BDB.R
\docType{data}
\name{PbP.BDB}
\alias{PbP.BDB}
\title{Play-by-play dataset - NBA 2017-2018}
\format{A data.frame with 37430 rows and 48 variables:
\describe{
\item{game_id}{ID}
\item{data_set}{Dataset}
\item{date}{date}
\item{a1}{a1}
\item{a2}{a2}
}}
\source{
\url{https://github.com/sndmrc/BasketballAnalyzeR}
}
\usage{
PbP.BDB
}
\description{
In this data frame, called PbP, cases (rows) are shots and variables (columns) are referred to
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
\keyword{datasets}
