% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PbPmanipulation.R
\name{PbPmanipulation}
\alias{PbPmanipulation}
\title{Manipulate play-by-play (PbP) dataset}
\usage{
PbPmanipulation(data, playTeam = "GSW")
}
\arguments{
\item{data}{PbP dataframe}

\item{playTeam}{Playing team}
}
\value{
A Pbp data frame
}
\description{
Manipulate play-by-play (PbP) dataset
}
