% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.networkdata.R
\name{plot.assistnet}
\alias{plot.assistnet}
\title{Create a network plot}
\usage{
\method{plot}{assistnet}(
  x,
  layout = "kamadakawai",
  layout.par = list(),
  edge.thr = 0,
  edge.col.lim = NULL,
  edge.col.lab = NULL,
  node.size = NULL,
  node.size.lab = NULL,
  node.col = NULL,
  node.col.lim = NULL,
  node.col.lab = NULL,
  node.pal = colorRampPalette(c("white", "blue", "red")),
  edge.pal = colorRampPalette(c("white", "blue", "red")),
  ...
)
}
\arguments{
\item{x}{An object of class 'networkdata'}

\item{layout}{Network vertex layout algorithms}

\item{layout.par}{Paramater for network vertex layout algorithms}

\item{edge.thr}{Threshold for edge values; values below the threshold are set to 0}

\item{edge.col.lim}{A numeric vector of length two providing limits of the scale for edge color}

\item{edge.col.lab}{Label for edge color legend}

\item{node.size}{A string with the variable name for node size}

\item{node.size.lab}{Label for node size legend}

\item{node.col}{A string with the variable name for node color}

\item{node.col.lim}{A numeric vector of length two providing limits of the scale for node color}

\item{node.col.lab}{Label for node color legend}

\item{node.pal}{Palette for node colors}

\item{edge.pal}{Palette for edge colors}

\item{...}{other graphical parameters}
}
\value{
A ggplot2 object
}
\description{
Create a network plot
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP.GSW <- subset(PbP, team=="GSW" & player!="")
out <- assistnet(PbP.GSW)
plot(out, layout="circle", edge.thr=30, node.col="FGM_ASTp", node.size="ASTPTS")
}
