% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.corr.analysis.R
\name{plot.corr.analysis}
\alias{plot.corr.analysis}
\title{Plot inequality}
\usage{
\method{plot}{corr.analysis}(x, title = NULL, ...)
}
\arguments{
\item{x}{A 'inequality' objects}

\item{title}{Plot title}

\item{...}{other graphical parameters}
}
\value{
A ggplot2 object
}
\description{
Plot inequality
}
\examples{
data <- data.frame(Pbox$PTS,Pbox$P3M,Pbox$P2M,
                   Pbox$OREB + Pbox$DREB,Pbox$AST,
                   Pbox$TOV,Pbox$STL,Pbox$BLK)/Pbox$MIN
names(data) <- c("PTS","P3M","P2M","REB","AST","TOV","STL","BLK")
data <- subset(data, Pbox$MIN >= 500)
out <- corr.analysis(data)
plot(out)
}
