% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.inequality.R
\name{plot.inequality}
\alias{plot.inequality}
\title{Plot inequality}
\usage{
\method{plot}{inequality}(x, title = NULL, ...)
}
\arguments{
\item{x}{A 'inequality' objects}

\item{title}{Plot title}

\item{...}{other graphical parameters}
}
\value{
A ggplot2 object
}
\description{
Plot inequality
}
\examples{
Pbox.BN <- subset(Pbox, Team=="Brooklyn Nets")
out <- inequality(Pbox.BN$PTS, nplayers=8)
print(out)
plot(out)
}
