% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MDSmap.R
\name{plot.MDSmap}
\alias{plot.MDSmap}
\title{Draw some two-dimensional plots for multidimensional scaling (MDS)}
\usage{
\method{plot}{MDSmap}(x, z.var = NULL, level.plot = TRUE,
  title = NULL, labels = NULL, repel_labels = FALSE,
  text_label = TRUE, subset = NULL, col.subset = "gray50",
  zoom = NULL, palette = NULL, contour = FALSE,
  ncol.arrange = NULL, ...)
}
\arguments{
\item{x}{An object of class 'MDSmap'}

\item{z.var}{The stratification variable (color)}

\item{level.plot}{If TRUE, draw a level plot, otherwise draw a scatter plot}

\item{title}{Plot title}

\item{labels}{Text labels for (x, y) points; only for single scatter plot}

\item{repel_labels}{If TRUE, draw text labels of not evidenced points using repelling}

\item{text_label}{If TRUE, draw a rectangle behind the evidenced text}

\item{subset}{Subset of points to be evidenced}

\item{col.subset}{Color for the subset of points}

\item{zoom}{X and Y axis range}

\item{palette}{Color palette}

\item{contour}{If TRUE, contour lines are plotted}

\item{ncol.arrange}{Number of columns when arranging multiple grobs on a page}

\item{...}{other graphical parameters}
}
\value{
A single ggplot2 plot or a list of ggplot2 plots
}
\description{
Draw some two-dimensional plots for multidimensional scaling (MDS)
}
\examples{
data <- data.frame(Pbox$PTS, Pbox$P3M, Pbox$P2M, Pbox$OREB + Pbox$DREB, Pbox$AST,
Pbox$TOV,Pbox$STL, Pbox$BLK)
names(data) <- c('PTS','P3M','P2M','REB','AST','TOV','STL','BLK')
selp <- which(Pbox$MIN >= 1500)
data <- data[selp,]
id <- Pbox$Player[selp]
mds <- MDSmap(data)
plot(mds, labels=id, z.var="P2M", level.plot=FALSE, palette=rainbow)
}
