% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.shotperformance.R
\name{plot.shotperformance}
\alias{plot.shotperformance}
\title{Plots a bubbleplot representing the data contained in the dataframe produced by the function 'shotperformance'}
\usage{
\method{plot}{shotperformance}(x, title = "Shooting performance", ...)
}
\arguments{
\item{x}{an object of class \code{ashotperformance} obtained using the shotperformance function}

\item{title}{character, plot title.}

\item{...}{other graphical parameters.}
}
\value{
A \code{ggplot2} object
}
\description{
Plots a bubbleplot representing the data contained in the dataframe produced by the function 'shotperformance'
}
\examples{
# Draw the plot for the performances on 2 point shots, when the high pressure situation is
# the one regarding shots taken when \code{shotclock} is between 0 and 2

PbP <- PbPmanipulation(PbP.BDB)
PbP <- scoredifference(PbP, team_name = "GSW", player_data=Pbox, team_data = Tadd)
PbP <- shotclock(PbP, sec_14_after_oreb = FALSE, team_data = Tadd)
players_perf <- shotperformance(PbP, shotclock_interval = c(0, 2),
                                player_data=Pbox, team_data = Tadd,
                                shot_type = "2P", teams = "GSW")
plot(players_perf)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.

P. Zuccolotto, M. Manisera and M. Sandri (2018) Big data analytics for modeling scoring probability in basketball: The effect of shooting under high pressure conditions. International Journal of Sports Science & Coaching.
}
\author{
Andrea Fox
}
