% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialprofile.R
\name{radialprofile}
\alias{radialprofile}
\title{Plot radial plot for player profiles}
\usage{
radialprofile(data, perc = FALSE, std = TRUE, title = NULL,
  ncol.arrange = NULL, min.mid.max = NULL)
}
\arguments{
\item{data}{A dataframe}

\item{perc}{perc}

\item{std}{std}

\item{title}{Title(s) for radial plots(s)}

\item{ncol.arrange}{The number of columns in the grid of arranged plots}

\item{min.mid.max}{A numerical vector with 3 elements: lower bound, middle dashed line, upper bound for radial axis}
}
\value{
A list of ggplot2 radial plots or a single ggplot2 plot of arranged radial plots
}
\description{
Plot radial plot for player profiles
}
\examples{
data("Pbox")
Pbox.PG <- Pbox[1:6,]
X <- data.frame(Pbox.PG$P2M, Pbox.PG$P3M, Pbox.PG$OREB+Pbox.PG$DREB,
                Pbox.PG$AST, Pbox.PG$TO)/Pbox.PG$MIN
names(X) <- c("P2M","P3M","REB","AST","TO")
radialprofile(data=X, ncol.arrange=3, title=Pbox.PG$Player)
}
