% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplereg.R
\name{simplereg}
\alias{simplereg}
\title{Simple regression}
\usage{
simplereg(x, y, type = "lin", sp = NULL)
}
\arguments{
\item{x}{data for x-axis}

\item{y}{data for y-axis}

\item{type}{type of regression: linear ('lin'), polynomial local fitting ('pol'), kernel smoothing ('ks')}

\item{sp}{parameter to control the degree of smoothing: span for loess and bandwidth for ksmooth (optional)}
}
\description{
Simple regression
}
