% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tadd.R
\docType{data}
\name{Tadd}
\alias{Tadd}
\title{Tadd dataset - NBA 2017-2018}
\format{A data frame with 30 rows and 23 variables:
\describe{
  \item{Team}{Teams}
  \item{Conference}{Conference}
  \item{Division}{Division}
  \item{Rank}{Rank}
  \item{Playoff}{Playoff}
}}
\source{
\url{https://github.com/sndmrc/BasketballAnalyzeR}
}
\usage{
Tadd
}
\description{
In this data frame cases (rows) are teams and variables (columns) are referred to achievements of the opponents in the different games
}
\keyword{datasets}
