% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistnet.R
\name{assistnet}
\alias{assistnet}
\title{Investigates the network of assists-shots in a team}
\usage{
assistnet(
  data,
  assist = "assist",
  player = "player",
  points = "points",
  event.type = "event_type"
)
}
\arguments{
\item{data}{A data frame whose rows are field shots and columns are variables to be specified in \code{assist}, \code{player}, \code{points}, \code{event.type} (see Details).}

\item{assist}{character, indicating the name of the variable with players who made the assists.}

\item{player}{character, indicating the name of the variable with players who made the shot.}

\item{points}{character, indicating the name of the variable with points.}

\item{event.type}{character, indicating the name of the variable with type of event (mandatory categories are \code{"miss"} for missed field shots and \code{"shot"} for field goals).}
}
\value{
A list with 3 elements, \code{assistTable} (a table), \code{nodeStats} (a dataframe), and a \code{assistNet} (a network object). See Details.
}
\description{
Investigates the network of assists-shots in a team
}
\details{
The \code{data} dataframe could also be a play-by-play dataset provided that rows corresponding to events different from field shots are not coded as \code{"shot"} in the \code{event.type} variable.

\code{assistTable} is the cross-table of assists made and received by the players.

\code{nodeStats} is a dataframe with the variables

- \code{FGM} (fields goals made),

- \code{FGM_AST} (field goals made thanks to a teammate's assist),

- \code{FGM_ASTp} (percentage of \code{FGM_AST} over \code{FGM}),

- \code{FGPTS} (points scored with field goals),

- \code{FGPTS_AST} (points scored thanks to a teammate's assist),

- \code{FGPTS_ASTp} (percentage of \code{FGPTS_AST} over \code{FGPTS}),

- \code{AST} (assists made),

- \code{ASTPTS} (point scored by assist's teammates).

\code{assistNet} is a object of class \code{network} that can be used for further network analysis with specific R packages.
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP.GSW <- subset(PbP, team=="GSW" & player!="")
out <- assistnet(PbP.GSW)
plot(out, layout="circle", edge.thr=30, node.col="FGM_ASTp", node.size="ASTPTS")
}
