% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barline.R
\name{barline}
\alias{barline}
\title{Create a graph with a barplot and a line plot}
\usage{
barline(data, id, bars, line, order.by = id, labels.bars = NULL,
  label.line = NULL, title = "")
}
\arguments{
\item{data}{A dataframe}

\item{id}{The name of the ID variable}

\item{bars}{Vector of names of the bar variables}

\item{line}{The name of the line variable}

\item{order.by}{The name of the variable used to order name players (on the x axis)}

\item{labels.bars}{Vector of labels for the bar variables}

\item{label.line}{The label for the line variable on the second axis (on the right)}

\item{title}{Title for the plot}
}
\value{
A ggplot2 object
}
\description{
Create a graph with a barplot and a line plot
}
\examples{
data("Pbox")
dts <- subset(Pbox, Team=="Houston Rockets" & MIN>=500)
barline(data=dts, id="Player", bars=c("P2p","P3p","FTp"),
        line="MIN", order.by="Player",
        labels.bars=c("2P","3P","FT"), title="Houston Rockets")
}
