% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplot.r
\name{densityplot}
\alias{densityplot}
\title{Shot density plot}
\usage{
densityplot(
  data,
  var,
  shot.type = "field",
  thresholds = NULL,
  best.scorer = FALSE,
  period.length = 12,
  bw = NULL,
  title = NULL
)
}
\arguments{
\item{data}{A play-by-play data frame with the following columns: \code{"ShotType"}, \code{"shot_distance"}, \code{"player"}, \code{"points"} and the column specified in \code{var}.}

\item{var}{A character string giving the numerical variable whose density to be estimated. Available options: "playlength", "periodTime", "totalTime", "shot_distance".}

\item{shot.type}{A character string giving the type of shots to be analyzed. Available options: "2P", "3P", "FT", field".}

\item{thresholds}{A numerical vector with two thresholds defining the range boundaries of the three regions.}

\item{best.scorer}{If TRUE, display the player who scored the higher number of points in the corresponding interval.}

\item{period.length}{period.length}

\item{bw}{A numerical value for the smoothing bandwidth of the kernel density estimator or a character string giving a rule to choose the bandwidth (see \link[stats]{density}).}

\item{title}{Plot title.}
}
\value{
A ggplot2 plot
}
\description{
Shot density plot
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
data.team  <- subset(PbP, team=="GSW" & result!="")
densityplot(data=data.team, shot.type="2P", var="playlength", best.scorer=TRUE)
data.opp <- subset(PbP, team!="GSW" & result!="")
densityplot(data=data.opp, shot.type="2P", var="shot_distance", best.scorer=TRUE)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
