% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedpts.R
\name{expectedpts}
\alias{expectedpts}
\title{Plot expected points of player shots as a function of the distance from the basket}
\usage{
expectedpts(
  data,
  players = NULL,
  bw = 10,
  palette = gg_color_hue,
  team = TRUE,
  col.team = "gray",
  col.hline = "black",
  xlab = NULL,
  title = NULL,
  legend = TRUE,
  var = "shot_distance"
)
}
\arguments{
\item{data}{a data frame whose rows are field shots and columns are variables to be specified in \code{var} and optionally in \code{players}.}

\item{players}{subset of players to be displayed}

\item{bw}{numeric, smoothing bandwidth of the kernel density estimator (see \code{\link[stats]{ksmooth}})}

\item{palette}{color palette}

\item{team}{logical; if \code{TRUE}, draw the scoring probability line for team}

\item{col.team}{character, color of the scoring probability line for team(default \code{"gray"})}

\item{col.hline}{character, color of the dashed horizontal line (default \code{"black"})}

\item{xlab}{character, x-axis label}

\item{title}{character, plot title}

\item{legend}{logical, if \code{TRUE}, color legend is displayed (only when \code{players} is not \code{NULL})}

\item{var}{character, numerical variable on the x-axis (default \code{"shot_distance"}).}
}
\value{
A \code{ggplot2} plot
}
\description{
Plot expected points of player shots as a function of the distance from the basket
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP.GSW <- subset(PbP, team=="GSW" & !is.na(shot_distance))
plrys <- c("Stephen Curry","Kevin Durant")
expectedpts(data=PbP.GSW, bw=10, players=plrys, col.team='dodgerblue',
        palette=colorRampPalette(c("gray","black")), col.hline="red")
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
