% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourfactors.R
\name{fourfactors}
\alias{fourfactors}
\title{Calculates possession, pace, ratings, and Four Factors}
\usage{
fourfactors(TEAM, OPP)
}
\arguments{
\item{TEAM}{A dataframe}

\item{OPP}{A dataframe}
}
\value{
A data frame with possession, pace, ratings and Four Factors for the analyzed teams and the opponents
}
\description{
Calculates possession, pace, ratings, and Four Factors
}
\examples{
selTeams <- c(2,6,10,11)
FF <- fourfactors(Tbox[selTeams,], Obox[selTeams,])
plot(FF)
}
