% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourfactors.R
\name{fourfactors}
\alias{fourfactors}
\title{Calculates possession, pace, ratings, and Four Factors}
\usage{
fourfactors(TEAM, OPP)
}
\arguments{
\item{TEAM}{a data frame (see Details).}

\item{OPP}{a data frame (see Details).}
}
\value{
An object of class \code{fourfactors}, i.e. a data frame with possession, pace, ratings and Four Factors for the analyzed teams and the opponents.
}
\description{
Calculates possession, pace, ratings, and Four Factors
}
\examples{
selTeams <- c(2,6,10,11)
FF <- fourfactors(Tbox[selTeams,], Obox[selTeams,])
plot(FF)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\seealso{
\code{\link{plot.fourfactors}}
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
