% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourfactors.R
\name{fourfactors}
\alias{fourfactors}
\title{Calculates Possession, pace, ratings, and Four Factors}
\usage{
fourfactors(TEAM, OPP, sel = 1:nrow(TEAM))
}
\arguments{
\item{TEAM}{A dataframe}

\item{OPP}{A dataframe}

\item{sel}{Vector of selected elements}
}
\value{
A data frame with possession, pace, ratings and Four Factors for the analyzed teams and the opponents
}
\description{
Calculates Possession, pace, ratings, and Four Factors
}
\examples{
data("Tbox")
data("Pbox")
selectedTeams <- c(2,6,10,11)
fourfactors(Tbox, Obox, sel=selectedTeams)
}
