% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustering.R
\name{hclustering}
\alias{hclustering}
\title{Agglomerative hierarchical clustering}
\usage{
hclustering(data, k = NULL, nclumax = 10, labels = NULL, linkage = "ward.D")
}
\arguments{
\item{data}{numeric  data frame}

\item{k}{integer, number of clusters}

\item{nclumax}{integer, maximum number of clusters (when \code{k=NULL})}

\item{labels}{character, row labels}

\item{linkage}{character, the agglomeration method to be used in \code{hclust} (see \code{method} in \link[stats]{hclust})}
}
\value{
A \code{hclustering} object, i.e. a list of 5 elements:

\itemize{
\item \code{k} (number of clusters),
}

\itemize{
\item \code{Subjects} (subjects' cluster identifiers),
}

\itemize{
\item \code{ClusterList} (clusters' composition),
}

\itemize{
\item \code{Profiles} (clusters' profiles, i.e. the average of the variables within clusters),
}

\itemize{
\item \code{Hclust} (an object of class \code{hclust}, see \code{\link[stats]{hclust}})
}
}
\description{
Agglomerative hierarchical clustering
}
\examples{
data <- data.frame(Pbox$PTS,Pbox$P3M,
                   Pbox$OREB + Pbox$DREB, Pbox$AST,
                   Pbox$TOV, Pbox$STL, Pbox$BLK,Pbox$PF)
names(data) <- c("PTS","P3M","REB","AST","TOV","STL","BLK","PF")
data <- subset(data, Pbox$MIN >= 1500)
ID <- Pbox$Player[Pbox$MIN >= 1500]
out <- hclustering(data, labels=ID, k=7)
plot(out)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\seealso{
\code{\link{plot.hclustering}}, \code{\link[stats]{hclust}}
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
