% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustering.R
\name{hclustering}
\alias{hclustering}
\title{Hierarchical clustering of NBA teams}
\usage{
hclustering(data, k = NULL, nclumax = 10, labels = NULL, linkage = "ward.D")
}
\arguments{
\item{data}{A dataframe}

\item{k}{Numer of clusters}

\item{nclumax}{Max number of clusters (when k=NULL)}

\item{labels}{labels}

\item{linkage}{linkage}
}
\value{
A 'hclustering' object, i.e. a list of ...(specificare)
}
\description{
Hierarchical clustering of NBA teams
}
\examples{
data <- data.frame(Pbox$PTS,Pbox$P3M,
                   Pbox$OREB + Pbox$DREB, Pbox$AST,
                   Pbox$TOV, Pbox$STL, Pbox$BLK,Pbox$PF)
names(data) <- c("PTS","P3M","REB","AST","TOV","STL","BLK","PF")
data <- subset(data, Pbox$MIN >= 1500)
ID <- Pbox$Player[Pbox$MIN >= 1500]
out <- hclustering(data, labels=ID, k=7)
plot(out)
}
