% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.assistnet.R
\name{is.assistnet}
\alias{is.assistnet}
\title{Reports whether x is a networkdata object}
\usage{
is.assistnet(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
returns TRUE if its argument is of class \code{networkdata} and FALSE otherwise
}
\description{
Reports whether x is a networkdata object
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP.GSW <- subset(PbP, team=="GSW" & player!="")
out <- assistnet(PbP.GSW)
is.assistnet(out)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\seealso{
\code{\link{assistnet}}
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
