% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.fourfactors.R
\name{is.fourfactors}
\alias{is.fourfactors}
\title{Reports whether x is a fourfactors object}
\usage{
is.fourfactors(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
returns TRUE if its argument is of class 'fourfactors' and FALSE otherwise
}
\description{
Reports whether x is a fourfactors object
}
\examples{
data("Tbox")
data("Pbox")
selTeams <- c(2,6,10,11)
out <- fourfactors(Tbox[selTeams,], Obox[selTeams,])
is.fourfactors(out)
}
