% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.fourfactors.R
\name{is.fourfactors}
\alias{is.fourfactors}
\title{Reports whether x is a 'fourfactors' object}
\usage{
is.fourfactors(x)
}
\arguments{
\item{x}{an object to test.}
}
\value{
Returns TRUE if its argument is of class \code{fourfactors} and FALSE otherwise.
}
\description{
Reports whether x is a 'fourfactors' object
}
\examples{
selTeams <- c(2,6,10,11)
out <- fourfactors(Tbox[selTeams,], Obox[selTeams,])
is.fourfactors(out)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\seealso{
\code{\link{fourfactors}}
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketball.analyzer.help@gmail.com})
}
