% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.hclust.R
\name{is.hclust}
\alias{is.hclust}
\title{Functions to check if an object is of class 'hclust'.}
\usage{
is.hclust(x)
}
\arguments{
\item{x}{any R object}
}
\value{
returns TRUE if its argument is of class 'hclust' and FALSE otherwise
}
\description{
Functions to check if an object is of class 'hclust'.
}
\examples{
data <- data.frame(Pbox$PTS,Pbox$P3M,
                   Pbox$OREB + Pbox$DREB, Pbox$AST,
                   Pbox$TOV, Pbox$STL, Pbox$BLK,Pbox$PF)
names(data) <- c("PTS","P3M","REB","AST","TOV","STL","BLK","PF")
data <- subset(data, Pbox$MIN >= 1500)
ID <- Pbox$Player[Pbox$MIN >= 1500]
out <- h.clustering(data, labels=ID, k=7)
is.hclust(out)
}
