% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.inequality.R
\name{is.inequality}
\alias{is.inequality}
\title{Functions to check if an object is of class 'inequality'.}
\usage{
is.inequality(x)
}
\arguments{
\item{x}{any R object}
}
\value{
returns TRUE if its argument is of class 'inequality' and FALSE otherwise
}
\description{
Functions to check if an object is of class 'inequality'.
}
\examples{
Pbox.BN <- subset(Pbox, Team=="Brooklyn Nets")
out <- inequality(Pbox.BN$PTS, npl=8)
is.inequality(out)
}
