% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.kclustering.R
\name{is.kclustering}
\alias{is.kclustering}
\title{Functions to check if an object is of class 'kclustering'.}
\usage{
is.kclustering(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
returns TRUE if its argument is of class \code{kclustering} and FALSE otherwise
}
\description{
Functions to check if an object is of class 'kclustering'.
}
\examples{
FF <- fourfactors(Tbox,Obox)
X <- with(FF, data.frame(OD.Rtg=ORtg/DRtg,
               F1.r=F1.Def/F1.Off, F2.r=F2.Off/F2.Def,
               F3.O=F3.Def, F3.D=F3.Off))
X$P3M <- Tbox$P3M
X$STL.r <- Tbox$STL/Obox$STL
kclu <- kclustering(X)
is.kclustering(kclu)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\seealso{
\code{\link{kclustering}}
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketball.analyzer.help@gmail.com})
}
