% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.MDSmap.R
\name{is.MDSmap}
\alias{is.MDSmap}
\title{Functions to check if an object is of class 'MDSmap'.}
\usage{
is.MDSmap(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
returns TRUE if its argument is of class \code{MDSmap} and FALSE otherwise
}
\description{
Functions to check if an object is of class 'MDSmap'.
}
\examples{
data <- subset(Pbox, MIN >= 1500)
data <- data.frame(data$PTS, data$P3M, data$P2M, data$OREB + data$DREB, data$AST,
                   data$TOV,data$STL, data$BLK)
mds <- MDSmap(data)
is.MDSmap(mds)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\seealso{
\code{\link{MDSmap}}
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketball.analyzer.help@gmail.com})
}
