% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.variability.R
\name{is.variability}
\alias{is.variability}
\title{Functions to check if an object is of class 'variability'.}
\usage{
is.variability(x)
}
\arguments{
\item{x}{any R object}
}
\value{
returns TRUE if its argument is of class 'variability' and FALSE otherwise
}
\description{
Functions to check if an object is of class 'variability'.
}
\examples{
Pbox.BC <- subset(Pbox, Team=="Oklahoma City Thunder" & MIN >= 500,
                    select=c("P2p","P3p","FTp","P2A","P3A","FTA"))
out <- variability(data=Pbox.BC, data.var=c("P2p","P3p","FTp"),
                   size.var=c("P2A","P3A","FTA"), weight=TRUE)
is.variability(out)
}
