% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kclustering.R
\name{kclustering}
\alias{kclustering}
\title{k-means clustering of NBA teams}
\usage{
kclustering(data, k = NULL, labels = NULL, nclumax = 10,
  nruns = 10)
}
\arguments{
\item{data}{A dataframe}

\item{k}{Numer of clusters}

\item{labels}{labels}

\item{nclumax}{Max number of clusters (when k=NULL)}

\item{nruns}{nruns}
}
\value{
A 'klust' object, i.e. a list of ...(specificare)
}
\description{
k-means clustering of NBA teams
}
\examples{
FF <- fourfactors(Tbox,Obox)
OD_Rtg <- FF$ORT/FF$DRT
F1_r <- FF$F1_O/FF$F1_T
F2_r <- FF$F2_T/FF$F2_O
F3_O <- FF$F3_O
F3_D <- FF$F3_T
P3M <- Tbox$P3M
STL_r <- Tbox$STL/Obox$STL
X <- data.frame(OD_Rtg,F1_r,F2_r,F3_O,F3_D,P3M,STL_r)
kclu1 <- kclustering(X)
plot(kclu1)
kclu2 <- kclustering(X, k=9)
plot(kclu2)
}
