% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kclustering.R
\name{kclustering}
\alias{kclustering}
\title{K-means  cluster analysis}
\usage{
kclustering(
  data,
  k = NULL,
  labels = NULL,
  nclumax = 10,
  nruns = 10,
  iter.max = 50,
  algorithm = "Hartigan-Wong"
)
}
\arguments{
\item{data}{numeric data frame}

\item{k}{integer, number of clusters}

\item{labels}{character, row labels}

\item{nclumax}{integer, maximum number of clusters (when \code{k=NULL}) used for calculating the explained variance as function of the number of clusters}

\item{nruns}{integer, run the k-means algorithm \code{nruns} times and chooses the best solution according to a maximum explained variance criterion}

\item{iter.max}{integer, maximum number of iterations allowed in k-means clustering (see \link[stats]{kmeans})}

\item{algorithm}{character, the algorithm used in k-means clustering (see \link[stats]{kmeans})}
}
\value{
A \code{klustering} object, i.e. a list of 4 elements:

\itemize{
\item \code{k} (number of clusters),
}

\itemize{
\item \code{Subjects} (subjects' cluster identifiers),
}

\itemize{
\item \code{ClusterList} (clusters' composition),
}

\itemize{
\item \code{Profiles} (clusters' profiles, i.e. the average of the variables within clusters)
}
}
\description{
K-means  cluster analysis
}
\examples{
FF <- fourfactors(Tbox,Obox)
OD.Rtg <- FF$ORtg/FF$DRtg
F1.r <- FF$F1.Def/FF$F1.Off
F2.r <- FF$F2.Off/FF$F2.Def
F3.Off <- FF$F3.Def
F3.Def <- FF$F3.Off
P3M <- Tbox$P3M
STL.r <- Tbox$STL/Obox$STL
X <- data.frame(OD.Rtg,F1.r,F2.r,F3.Off,F3.Def,P3M,STL.r)
kclu1 <- kclustering(X)
plot(kclu1)
kclu2 <- kclustering(X, k=9)
plot(kclu2)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\seealso{
\code{\link{plot.kclustering}}, \code{\link[stats]{kmeans}}
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
