% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kclustering.R
\name{kclustering}
\alias{kclustering}
\title{k-means clustering of NBA teams}
\usage{
kclustering(data, k = NULL, labels = NULL, nclumax = 10,
  nruns = 10, iter.max = 50, algorithm = "Hartigan-Wong")
}
\arguments{
\item{data}{A dataframe}

\item{k}{Numer of clusters}

\item{labels}{labels}

\item{nclumax}{Maximum number of clusters (when k=NULL)}

\item{nruns}{nruns}

\item{iter.max}{Maximum number of iterations allowed in k-means clustering}

\item{algorithm}{Algorithm used in k-means clustering (see \link[stats]{kmeans})}
}
\value{
A 'klust' object, i.e. a list of ...(specificare)
}
\description{
k-means clustering of NBA teams
}
\examples{
FF <- fourfactors(Tbox,Obox)
OD.Rtg <- FF$ORtg/FF$DRtg
F1.r <- FF$F1.Def/FF$F1.Off
F2.r <- FF$F2.Off/FF$F2.Def
F3.Off <- FF$F3.Def
F3.Def <- FF$F3.Off
P3M <- Tbox$P3M
STL.r <- Tbox$STL/Obox$STL
X <- data.frame(OD.Rtg,F1.r,F2.r,F3.Off,F3.Def,P3M,STL.r)
kclu1 <- kclustering(X)
plot(kclu1)
kclu2 <- kclustering(X, k=9)
plot(kclu2)
}
