% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSmap.R
\name{MDSmap}
\alias{MDSmap}
\title{Create some two-dimensional plots for multidimensional scaling (MDS)}
\usage{
MDSmap(data, std = TRUE)
}
\arguments{
\item{data}{A dataframe to use for plot}

\item{std}{If TRUE, columns of 'data' are standardized}
}
\value{
An object of class 'MDSmap': a list with 4 objects: points, stress, data, std
}
\description{
Create some two-dimensional plots for multidimensional scaling (MDS)
}
\examples{
data <- data.frame(Pbox$PTS, Pbox$P3M, Pbox$P2M, Pbox$OREB + Pbox$DREB, Pbox$AST,
Pbox$TOV,Pbox$STL, Pbox$BLK)
names(data) <- c('PTS','P3M','P2M','REB','AST','TOV','STL','BLK')
selp <- which(Pbox$MIN >= 1500)
data <- data[selp,]
id <- Pbox$Player[selp]
mds <- MDSmap(data)
plot(mds, labels=id, z.var="P2M", level.plot=FALSE, palette=rainbow)
}
