% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSmap.R
\name{MDSmap}
\alias{MDSmap}
\title{Create some two-dimensional plots for multidimensional scaling (MDS)}
\usage{
MDSmap(data, labels = NULL, subset = NULL, zoom = NULL, var = NULL,
  title = NULL, colors = NULL, contour = FALSE, col.palette = NULL,
  arrange = FALSE, ncol.arrange = NULL)
}
\arguments{
\item{data}{A dataframe to use for plot}

\item{labels}{labels}

\item{subset}{subset}

\item{zoom}{zoom}

\item{var}{var}

\item{title}{title}

\item{colors}{colors}

\item{contour}{contour}

\item{col.palette}{Colour palette}

\item{arrange}{A logical value. If TRUE radial plots are arranged in a single plot}

\item{ncol.arrange}{The number of columns in the grid of arranged plots}
}
\value{
A list with two objects: a list of plots (named 'listPlots') and a dataframe with MDS coordinates
}
\description{
Create some two-dimensional plots for multidimensional scaling (MDS)
}
\examples{
data("Pbox")
dts <- subset(Pbox, Team=="Houston Rockets" & MIN>=500)
barline(data=dts, id="Player", bars=c("P2p","P3p","FTp"),
        line="MIN", order.by="Player",
        labels.bars=c("2P","3P","FT"), title="Houston Rockets")
}
