% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pbox.R
\docType{data}
\name{Pbox}
\alias{Pbox}
\title{Players box scores dataset - NBA 2017-2018}
\format{A data.frame with 605 rows and 22 variables:
\describe{
\item{Team}{Analyzed team, character}
\item{Player}{Analyzed player, character}
\item{GP}{Games Played, numeric}
\item{MIN}{Minutes Played, numeric}
\item{PTS}{Points Made, numeric}
}}
\usage{
Pbox
}
\description{
In this data frame, called Pbox, cases (rows) are players and variables (columns) are referred to the individual achievements in the different games
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
\keyword{datasets}
