% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PbPmanipulation.R
\name{PbPmanipulation}
\alias{PbPmanipulation}
\title{Manipulate play-by-play (PbP) dataset}
\usage{
PbPmanipulation(data)
}
\arguments{
\item{data}{PbP dataframe}
}
\value{
A Pbp data frame
}
\description{
Manipulate play-by-play (PbP) dataset
}
