% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PbPmanipulation.R
\name{PbPmanipulation}
\alias{PbPmanipulation}
\title{Adapts the standard file supplied by BigDataBall to the format required by BasketballAnalyzeR}
\usage{
PbPmanipulation(data)
}
\arguments{
\item{data}{a PbP data frame supplied by \href{https://www.bigdataball.com/}{BigDataBall}}
}
\value{
A PbP data frame
}
\description{
Adapts the standard file supplied by BigDataBall to the format required by BasketballAnalyzeR
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\seealso{
\code{\link{PbP.BDB}}
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketball.analyzer.help@gmail.com})
}
