% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fourfactors.R
\name{plot.fourfactors}
\alias{plot.fourfactors}
\title{Plot possession, pace, ratings, and Four Factors}
\usage{
\method{plot}{fourfactors}(x, title = NULL, ...)
}
\arguments{
\item{x}{A data.frame with calculated possession, pace, ratings, and Four Factors}

\item{title}{Plot title}

\item{...}{other graphical parameters}
}
\value{
A list of four ggplot2 plots
}
\description{
Plot possession, pace, ratings, and Four Factors
}
\examples{
data("Tbox")
data("Pbox")
selectedTeams <- c(2,6,10,11)
FF <- fourfactors(Tbox, Obox, sel=selectedTeams)
plot(FF)
}
