% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hclustering.R
\name{plot.hclustering}
\alias{plot.hclustering}
\title{Plot hierarchical clustering of NBA teams}
\usage{
\method{plot}{hclustering}(x, title = NULL, profiles = FALSE,
  ncol.arrange = NULL, circlize = FALSE, horiz = TRUE,
  cex.labels = 0.7, colored.labels = TRUE, colored.branches = FALSE,
  rect = FALSE, lower.rect = NULL, min.mid.max = NULL, ...)
}
\arguments{
\item{x}{An object of class 'hclustering'}

\item{title}{Plot title}

\item{profiles}{Plot a radial plots of cluster mean profiles}

\item{ncol.arrange}{Number of columns when arranging multiple grobs on a page}

\item{circlize}{Plot a circular dendrogram}

\item{horiz}{Plot an horizontal dendrogram (only for non circular dendrograms)}

\item{cex.labels}{The magnification to be used for dendrogram labels}

\item{colored.labels}{Assign different colors to labels of different clusters}

\item{colored.branches}{Assign different colors to dendrogram branches of different clusters}

\item{rect}{Draw rectangles around the branches of a dendrogram in order to highlight the corresponding clusters}

\item{lower.rect}{A (scalar) value of how low should the lower part of the rect be}

\item{min.mid.max}{A numerical vector with 3 elements: lower bound, middle dashed line, upper bound for radial axis}

\item{...}{other graphical parameters}
}
\value{
A single ggplot2 plot or (when \code{profile=FALSE}) a list of (radial) plots (when \code{profile=TRUE}))
}
\description{
Plot hierarchical clustering of NBA teams
}
\examples{
data <- data.frame(Pbox$PTS,Pbox$P3M,
                   Pbox$OREB + Pbox$DREB, Pbox$AST,
                   Pbox$TOV, Pbox$STL, Pbox$BLK,Pbox$PF)
names(data) <- c("PTS","P3M","REB","AST","TOV","STL","BLK","PF")
data <- subset(data, Pbox$MIN >= 1500)
ID <- Pbox$Player[Pbox$MIN >= 1500]
out <- hclustering(data, labels=ID, k=7)
plot(out)
}
