% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.kclust.R
\name{plot.kclust}
\alias{plot.kclust}
\title{Plot k-means clustering of NBA teams}
\usage{
\method{plot}{kclust}(x, title = NULL, ncol.arrange = NULL, ...)
}
\arguments{
\item{x}{An object of class 'kclust'}

\item{title}{Plot title}

\item{ncol.arrange}{Number of columns when arranging multiple grobs on a page}

\item{...}{other graphical parameters}
}
\value{
A list of ggplot2 radial plots
}
\description{
Plot k-means clustering of NBA teams
}
\examples{
FF <- fourfactors(Tbox,Obox)
OD_Rtg <- FF$ORtg/FF$DRtg
F1_r <- FF$F1.D/FF$F1.O
F2_r <- FF$F2.O/FF$F2.D
F3_O <- FF$F3.D
F3_D <- FF$F3.O
P3M <- Tbox$P3M
STL_r <- Tbox$STL/Obox$STL
X <- data.frame(OD_Rtg,F1_r,F2_r,F3_O,F3_D,P3M,STL_r)
kclu1 <- kclustering(X)
plot(kclu1)
kclu2 <- kclustering(X, k=9)
plot(kclu2)
}
