% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.kclustering.R
\name{plot.kclustering}
\alias{plot.kclustering}
\title{Plot k-means clustering of NBA teams}
\usage{
\method{plot}{kclustering}(x, title = NULL, ncol.arrange = NULL, min.mid.max = NULL, ...)
}
\arguments{
\item{x}{An object of class 'kclust'}

\item{title}{Plot title}

\item{ncol.arrange}{Number of columns when arranging multiple grobs on a page}

\item{min.mid.max}{A numerical vector with 3 elements: lower bound, middle dashed line, upper bound for radial axis}

\item{...}{other graphical parameters}
}
\value{
A single plot or a list of ggplot2 radial plots
}
\description{
Plot k-means clustering of NBA teams
}
\examples{
FF <- fourfactors(Tbox,Obox)
OD.Rtg <- FF$ORtg/FF$DRtg
F1.r <- FF$F1.D/FF$F1.O
F2.r <- FF$F2.O/FF$F2.D
F3.O <- FF$F3.D
F3.D <- FF$F3.O
P3M <- Tbox$P3M
STL.r <- Tbox$STL/Obox$STL
X <- data.frame(OD.Rtg,F1.r,F2.r,F3.O,F3.D,P3M,STL.r)
kclu1 <- kclustering(X)
plot(kclu1)
kclu2 <- kclustering(X, k=9)
plot(kclu2)
}
