% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simplereg.R
\name{plot.simplereg}
\alias{plot.simplereg}
\title{Plot simple regression}
\usage{
\method{plot}{simplereg}(x, labels = NULL, subset = NULL, Lx = 0.01,
  Ux = 0.99, Ly = 0.01, Uy = 0.99, title = NULL, xtitle = NULL,
  ytitle = NULL, repel = TRUE, ...)
}
\arguments{
\item{x}{an object of class 'simplereg'}

\item{labels}{labels for the subjects (optional)}

\item{subset}{subset of subjects to be highlighted in the graph  (optional) - or 'quant' to highligh the upper and lower quantiles}

\item{Lx}{active in case of subset='auto'; specify the lower quantiles of x to display - default = 0.01}

\item{Ux}{active in case of subset='auto'; specify the upper quantiles of x to display - default = 0.99}

\item{Ly}{active in case of subset='auto'; specify the lower quantiles of y to display - default = 0.01}

\item{Uy}{active in case of subset='auto'; specify the upper quantiles of y to display - default = 0.99}

\item{title}{title}

\item{xtitle}{label of x-axis}

\item{ytitle}{label of y-axis}

\item{repel}{text labels repel away from each other}

\item{...}{other graphical parameters}
}
\description{
Plot simple regression
}
\examples{
Pbox.sel <- subset(Pbox, MIN >= 500)
X <- Pbox.sel$AST/Pbox.sel$MIN
Y <- Pbox.sel$TOV/Pbox.sel$MIN
Pl <- Pbox.sel$Player
mod <- simplereg(x=X, y=Y, type="lin")
plot(mod)
}
