% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.simplereg.R
\name{plot.simplereg}
\alias{plot.simplereg}
\title{Plot simple regression}
\usage{
\method{plot}{simplereg}(x, labels = NULL, subset = NULL, Lx = 0.01,
  Ux = 0.99, Ly = 0.01, Uy = 0.99, title = "Simple regression",
  xtitle = NULL, ytitle = NULL, repel = TRUE, ...)
}
\arguments{
\item{x}{An object of class 'simplereg'}

\item{labels}{Labels for the subjects (optional)}

\item{subset}{Subset of subjects to be highlighted in the graph  (optional) - or 'quant' to highligh the upper and lower quantiles}

\item{Lx}{Active in case of subset='auto'; specify the lower quantiles of x to display - default = 0.01}

\item{Ux}{Active in case of subset='auto'; specify the upper quantiles of x to display - default = 0.99}

\item{Ly}{Active in case of subset='auto'; specify the lower quantiles of y to display - default = 0.01}

\item{Uy}{Active in case of subset='auto'; specify the upper quantiles of y to display - default = 0.99}

\item{title}{Plot title}

\item{xtitle}{Label of x-axis}

\item{ytitle}{Label of y-axis}

\item{repel}{Text labels repel away from each other}

\item{...}{Other graphical parameters}
}
\description{
Plot simple regression
}
\examples{
Pbox.sel <- subset(Pbox, MIN >= 500)
X <- Pbox.sel$AST/Pbox.sel$MIN
Y <- Pbox.sel$TOV/Pbox.sel$MIN
Pl <- Pbox.sel$Player
mod <- simplereg(x=X, y=Y, type="lin")
plot(mod)
}
