% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.variability.R
\name{plot.variability}
\alias{plot.variability}
\title{Plot variability}
\usage{
\method{plot}{variability}(
  x,
  title = "Variability diagram",
  ylim = NULL,
  ylab = NULL,
  size.lim = NULL,
  max.circle = 25,
  n.circle = 4,
  leg.brk = NULL,
  leg.pos = "right",
  leg.just = "left",
  leg.nrow = NULL,
  leg.title = NULL,
  leg.title.pos = "top",
  ...
)
}
\arguments{
\item{x}{A 'variability' objects}

\item{title}{Plot title}

\item{ylim}{Set y-axis limits}

\item{ylab}{Set y-axis legend}

\item{size.lim}{Set limits of 'size' legend}

\item{max.circle}{Maximum size of the 'size' plotting symbol (circles)}

\item{n.circle}{Set a sequence of about n.circle+1 equally spaced 'round' values which cover the range of the values in 'size'}

\item{leg.brk}{Breaks for 'size' legend}

\item{leg.pos}{The position of legend ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{leg.just}{Anchor point for positioning legend inside plot ("center" or two-element numeric vector) or the justification according to the plot area when positioned outside the plot}

\item{leg.nrow}{The desired number of rows of 'size' legend}

\item{leg.title}{Title of 'size' legend}

\item{leg.title.pos}{A character string indicating the position of a title. One of "top" (default for a vertical legend), "bottom", "left" (default for a horizontal legend), or "right."}

\item{...}{other graphical parameters}
}
\value{
A ggplot2 object
}
\description{
Plot variability
}
\examples{
Pbox.BC <- subset(Pbox, Team=="Oklahoma City Thunder" & MIN >= 500,
                    select=c("P2p","P3p","FTp","P2A","P3A","FTA"))
out <- variability(data=Pbox.BC, data.var=c("P2p","P3p","FTp"),
                   size.var=c("P2A","P3A","FTA"), weight=TRUE)
plot(out, leg.brk=c(10,25,50,100,500,1000), max.circle=30)
}
