% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.variability.R
\name{plot.variability}
\alias{plot.variability}
\title{Plot variability}
\usage{
\method{plot}{variability}(x, title = NULL, ...)
}
\arguments{
\item{x}{A 'variability' objects}

\item{title}{Plot title}

\item{...}{other graphical parameters}
}
\value{
A ggplot2 object
}
\description{
Plot variability
}
\examples{
Pbox.BC <- subset(Pbox, Team=="Oklahoma City Thunder" & MIN >= 500,
                    select=c("P2p","P3p","FTp","P2A","P3A","FTA"))
out <- variability(data=Pbox.BC, data.var=c("P2p","P3p","FTp"),
                   size.var=c("P2A","P3A","FTA"), weight=TRUE)
plot(out)
}
