% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialprofile.R
\name{radialprofile}
\alias{radialprofile}
\title{Plot radial plot for player profiles}
\usage{
radialprofile(data, group, perc = FALSE, std = TRUE, title = NULL,
  arrange = FALSE, ncol.arrange = NULL)
}
\arguments{
\item{data}{A dataframe}

\item{group}{A string containing the name of the variable with players' names}

\item{perc}{perc}

\item{std}{std}

\item{title}{Plot title}

\item{arrange}{A logical value. If TRUE radial plots are arranged in a single plot}

\item{ncol.arrange}{The number of columns in the grid of arranged plots}
}
\value{
A list of ggplot2 radial plots or a single ggplot2 plot of arranged radial plots
}
\description{
Plot radial plot for player profiles
}
\examples{
data("Pbox")
Pbox.PG <- Pbox[1:15,]
X <- data.frame(Pbox.PG$P2M, Pbox.PG$P3M, Pbox.PG$OREB+Pbox.PG$DREB,
                Pbox.PG$AST, Pbox.PG$TO)/Pbox.PG$MIN
X <- cbind(Pbox.PG$Player,X)
names(X) <- c("ID","P2M","P3M","REB","AST","TO")
p <- radialprofile(data=X, group="ID", arrange=TRUE, ncol.arrange=5)
library(grid)
grid.draw(p)
}
