% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatter plot and matrix of scatter plot}
\usage{
scatterplot(data, data.var, z.var = NULL, palette = NULL,
  labels = NULL, repel_labels = FALSE, text_label = TRUE,
  subset = NULL, col.subset = "gray50", zoom = NULL, title = NULL,
  legend = "yes", upper = list(continuous = "cor", combo =
  "box_no_facet", discrete = "facetbar", na = "na"),
  lower = list(continuous = "points", combo = "facethist", discrete =
  "facetbar", na = "na"), diag = list(continuous = "densityDiag",
  discrete = "barDiag", na = "naDiag"))
}
\arguments{
\item{data}{A dataframe}

\item{data.var}{A vector of characaters with name of variables used on the axes of scatter plot(s)}

\item{z.var}{The stratification variable (color)}

\item{palette}{Color palette}

\item{labels}{Text labels for (x, y) points; only for single scatter plot}

\item{repel_labels}{If TRUE, draw text labels of not evidenced points using repelling}

\item{text_label}{If TRUE, draw a rectangle behind the evidenced text}

\item{subset}{Subset of points to be evidenced; only for single scatter plot}

\item{col.subset}{Color for the subset of points; only for single scatter plot}

\item{zoom}{X and Y axis range; only for single scatter plot}

\item{title}{Plot title}

\item{legend}{Set legend='no' to remove legend}

\item{upper}{See help for GGally::ggpairs}

\item{lower}{See help for GGally::ggpairs}

\item{diag}{See help for GGally::ggpairs}
}
\value{
A ggplot2 plot
}
\description{
Scatter plot and matrix of scatter plot
}
\examples{
# Single scatter plot
Pbox.sel <- subset(Pbox, MIN>= 500)
X <- data.frame(AST=Pbox.sel$AST/Pbox.sel$MIN,TOV=Pbox.sel$TOV/Pbox.sel$MIN)
X$PTSpm <- Pbox.sel$PTS/Pbox.sel$MIN
mypal <- colorRampPalette(c("blue","yellow","red"))
scatterplot(X, data.var=c("AST","TOV"), z.var="PTSpm", labels=1:nrow(X), palette=mypal)
# Matrix of scatter plots
data <- Pbox[1:50, c("PTS","P3M","P2M","OREB","Team")]
scatterplot(data, data.var=1:4, z.var="Team")
}
