% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatter plot and matrix of scatter plot}
\usage{
scatterplot(data, z = NULL, z.name = NULL, palette = NULL,
  labels = NULL, text_label = TRUE, subset = NULL,
  col.subset = "gray50", zoom = NULL, title = NULL,
  upper = list(continuous = "cor", combo = "box_no_facet", discrete =
  "facetbar", na = "na"), lower = list(continuous = "points", combo =
  "facethist", discrete = "facetbar", na = "na"), diag = list(continuous
  = "densityDiag", discrete = "barDiag", na = "naDiag"))
}
\arguments{
\item{data}{A dataframe}

\item{z}{Stratification variable}

\item{z.name}{Name of stratification variable}

\item{palette}{Color palette}

\item{labels}{Text labels for (x, y) points; only for single scatter plot}

\item{text_label}{If TRUE, draw a rectangle behind the evidenced text}

\item{subset}{Subset of points to be evidenced; only for single scatter plot}

\item{col.subset}{Color for the subset of points; only for single scatter plot}

\item{zoom}{X and Y axis range; only for single scatter plot}

\item{title}{Plot title}

\item{upper}{See help for GGally::ggpairs}

\item{lower}{See help for GGally::ggpairs}

\item{diag}{See help for GGally::ggpairs}
}
\value{
A ggplot2 plot
}
\description{
Scatter plot and matrix of scatter plot
}
\examples{
# Single scatter plot
data <- Pbox[, c("PTS","P3M")]
scatterplot(data, labels=as.character(Pbox$Player), subset=1:3)
# Matrix of scatter plots
data <- Pbox[, c("PTS","P3M","P2M","OREB")]
scatterplot(data, z=Pbox$Team)
}
