% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringprob.R
\name{scoringprob}
\alias{scoringprob}
\title{Plots scoring probability of shots as a function of a given variable}
\usage{
scoringprob(
  data,
  var,
  shot.type,
  players = NULL,
  bw = 20,
  period.length = 12,
  title = NULL,
  palette = gg_color_hue,
  team = TRUE,
  col.team = "dodgerblue",
  legend = TRUE
)
}
\arguments{
\item{data}{a play-by-play data frame.}

\item{var}{character, the name of the numerical variable according to which the scoring probability is estimated. Available options: \code{"playlength"}, \code{"periodTime"}, \code{"totalTime"}, \code{"shot_distance"}.}

\item{shot.type}{character, the type of shots to be analyzed; available options: \code{"2P"}, \code{"3P"}, \code{"FT"}, \code{"field"}.}

\item{players}{subset of players to be displayed (optional; it can be used only if the \code{player} column is present in \code{data}).}

\item{bw}{numeric, the smoothing bandwidth of the kernel density estimator (see \link[stats]{ksmooth}).}

\item{period.length}{numeric, the length of a quarter in minutes (default: 12 minutes as in NBA).}

\item{title}{character, plot title.}

\item{palette}{color palette.}

\item{team}{character; if \code{TRUE} draws the scoring probability for all the shots in data.}

\item{col.team}{character, color of the scoring probability line for all the shots in data.}

\item{legend}{character; if \code{TRUE}, color legend is displayed (only when \code{players} is not \code{NULL}).}
}
\value{
A \code{ggplot2} plot
}
\description{
Plots scoring probability of shots as a function of a given variable
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP.GSW <- subset(PbP, team=="GSW" & result!="")
players <- c("Kevin Durant","Draymond Green","Klay Thompson")
p <- scoringprob(data=PbP.GSW, shot.type="2P", players=players,
                 var="shot_distance", col.team="gray")
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
