% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoringprob.R
\name{scoringprob}
\alias{scoringprob}
\title{Scoring probability plot}
\usage{
scoringprob(data, var, shot.type, players = NULL, bw = 20,
  period.length = 12, title = NULL, palette = NULL,
  col.team = "steelblue1", legend = TRUE)
}
\arguments{
\item{data}{A play-by-play data frame}

\item{var}{A character string giving the numerical variable whose density to be estimated. Available options: "playlength", "periodTime", "totalTime", "shot_distance".}

\item{shot.type}{A character string giving the type of shots to be analyzed. Available options: "2P", "3P", "field"}

\item{players}{Subset of players to be displayed}

\item{bw}{A numerical value for the smoothing bandwidth of the kernel density estimator or a character string giving a rule to choose the bandwidth (see \link[stats]{density})}

\item{period.length}{period.length}

\item{title}{Plot title}

\item{palette}{Color palette}

\item{col.team}{Color of the scoring probability line for team}

\item{legend}{If TRUE, color legend is displayed (only when 'players' is not NULL)}
}
\value{
A ggplot2 plot
}
\description{
Scoring probability plot
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
PbP.GSW <- subset(PbP, team=="GSW" & result!="")
players <- c("Kevin Durant","Draymond Green","Klay Thompson")
p <- scoringprob(data=PbP.GSW, shot.type="2P", players=players,
                 var="shot_distance", col.team="gray")
}
