% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotchart.R
\name{shotchart}
\alias{shotchart}
\title{Plot shortcharts}
\usage{
shotchart(data, x = NULL, y = NULL, z = NULL, result = NULL,
  type = NULL, scatter = FALSE, num.sect = 7, n = 1000,
  col.limits = c(NA, NA), courtline.col = "black", bg.col = "white",
  legend = TRUE, pt.col = "black", pt.alpha = 0.5, nbins = 25,
  palette = "mixed")
}
\arguments{
\item{data}{A data frame}

\item{x}{Variable name for the x coordinate}

\item{y}{Variable name for the y coordinate}

\item{z}{Name of 'data' variable to be used as a third variable}

\item{result}{Name of 'data' variable containing the type of shot; this variable is used to generate shot statistics; allowed categories: 'made' and 'missed'}

\item{type}{A string with the plot type: 'sectors', 'density-polygon', 'density-raster', 'density-hexbin'}

\item{scatter}{Logical. If TRUE a scatter plot was added to the plot}

\item{num.sect}{Number of sectors in shotchart plot split by sectors}

\item{n}{Number of point used when drawing arcs in shotchart plot split by sectors}

\item{col.limits}{Vector of limits c(min, max) for the gratient color scale}

\item{courtline.col}{Color of court lines}

\item{bg.col}{Background color}

\item{legend}{If TRUE a legend is plotted}

\item{pt.col}{Color of points in the scatter plot}

\item{pt.alpha}{Transparency of points in the scatter plot}

\item{nbins}{Number of bins in the 'density-hexbin' plot}

\item{palette}{Palette color for the scatter plot: 'main', 'cool', 'hot', 'mixed', 'grey'}
}
\value{
A ggplot2 object
}
\description{
Plot shortcharts
}
\examples{
PbP <- PbPmanipulation(PbP.BDB)
subdata <- subset(PbP, player=="Kevin Durant")
subdata$xx <- subdata$original_x/10
subdata$yy <- subdata$original_y/10-42
# Shotchart with colored sectors and statistics for a 3rd variable
shotchart(data=subdata, x="xx", y="yy", z="playlength", result="result",
          type="sectors",  num.sect=7)
# Shotchart with scatter plot and a stratification variable
shotchart(data=subdata, x="xx", y="yy", z="result",
          type=NULL, scatter=TRUE, palette="main")
}
