% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TOPboxes.r
\name{TOPboxes}
\alias{TOPboxes}
\title{Calculate Tbox, Obox and Pbox}
\usage{
TOPboxes(data, team)
}
\arguments{
\item{data}{a play-by-play  data frame}

\item{team}{character, team}
}
\value{
A list with the following elements

\itemize{
\item \code{Tbox}, (completare descrizione)
}

\itemize{
\item \code{Obox}, (completare descrizione)
}

\itemize{
\item \code{Pbox}, (completare descrizione)
}
}
\description{
Calculate Tbox, Obox and Pbox
}
\examples{
library(operators)
library(dplyr)
PbP <- PbPmanipulation(PbP.BDB)
PbP <- PbP \%>\%
  mutate(oreb = type \%~\% "rebound offensive",
         dreb = type \%~\% "rebound defensive",
                turnover = event_type=="turnover",
                PF = (event_type == "foul") & !(type \%~\% "technical") ) \%>\%
        mutate(across(c(player, assist, steal, block, h1:h5, a1:a5), as.character)) \%>\%
 as.data.frame()
out <- TOPboxes(PbP, team="GSW")
}
\references{
P. Zuccolotto and M. Manisera (2020) Basketball Data Science: With Applications in R. CRC Press.
}
\seealso{
\code{\link{PbPmanipulation}}
}
\author{
Marco Sandri, Paola Zuccolotto, Marica Manisera (\email{basketballanalyzer.help@unibs.it})
}
