% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variability.R
\name{variability}
\alias{variability}
\title{Variability analysis}
\usage{
variability(data, data.var, size.var, weight = FALSE)
}
\arguments{
\item{data}{A dataframe}

\item{data.var}{A vector of variable names or of column numbers defining variables whose variability will be analyzed by 'variability'}

\item{size.var}{A vector of variable names or of column numbers defining variables used for bubble sizes}

\item{weight}{If TRUE, calculates weighted variation coefficients}
}
\value{
A list with the following elements: ranges, standard deviations, variation coefficients, and two dataframes (data, size)
}
\description{
Variability analysis
}
\examples{
Pbox.BC <- subset(Pbox, Team=="Oklahoma City Thunder" & MIN >= 500,
                    select=c("P2p","P3p","FTp","P2A","P3A","FTA"))
list_variability <- variability(data=Pbox.BC, data.var=c("P2p","P3p","FTp"),
                                size.var=c("P2A","P3A","FTA"), weight=TRUE)
print(list_variability)
plot(list_variability, leg.brk=c(10,25,50,100,500,1000), max.circle=30)
}
